:card_file_box:(migrations):  creating and modifying migrations to adapt the table structure

+++ database/migrations/2025_09_11_212339_add_reactivation_users.php
@@ -13,9 +13,9 @@ return new class extends Migration
     {
         Schema::table('users', function (Blueprint $table) {
             $table->timestamp('reactivation_requested_at')->nullable()->after('actif');
-            $table->unsignedInteger('reactivation_requested_by')->nullable()->after('reactivation_requested_at');
+            $table->unsignedBigInteger('reactivation_requested_by')->nullable()->after('reactivation_requested_at');
             $table->timestamp('reactivation_approved_at')->nullable()->after('reactivation_requested_by');
-            $table->unsignedInteger('reactivation_approved_by')->nullable()->after('reactivation_approved_at');
+            $table->unsignedBigInteger('reactivation_approved_by')->nullable()->after('reactivation_approved_at');
 
             $table->foreign('reactivation_requested_by')->references('id')->on('users')->nullOnDelete();
             $table->foreign('reactivation_approved_by')->references('id')->on('users')->nullOnDelete();
